angular restructure
