+++ src/pages/integration-builder/builder/frameworks/angular/stepContent/buildingApp.mdx
